Added possibility to dequeue jobs (by @xezpeleta) and limited policies and scripts management for not-admin users (by @antton).

+++ src/Binovo/ElkarBackupBundle/Command/StopJobCommand.php
@@ -63,13 +63,24 @@
                                           $tmp,
                                           $clientId,
                                           $jobId);
-     if (file_exists($lockfile)) {
-       $command1 = shell_exec(sprintf("kill -TERM $(cat '%s')", $lockfile));
-       $command2 = shell_exec("killall rsync");
-       $this->info('Job backup aborted successfully', array(), $context);
-     } else {
-       $this->warn('Cannot abort job backup: not running', array(), $context);
-     }
+
+    if ($job->getStatus() == "ABORTED"){
+        $this->info('Job previously aborted by tick command', array(), $context);
+    } else {
+        if (file_exists($lockfile)) {
+           $command1 = shell_exec(sprintf("kill -TERM $(cat '%s')", $lockfile));
+           $command2 = shell_exec("killall rsync");
+           $this->info('Job backup aborted successfully', array(), $context);
+           $job->setStatus('ABORTED');
+           $context = array('link'   => $this->generateJobRoute($jobId, $clientId),
+                            'source' => Globals::STATUS_REPORT);
+           $job->setStatus('ABORTED');
+        } else {
+           $this->warn('Cannot abort job backup: not running', array(), $context);
+        }
+    }
+
+    return true;
    }
 
    protected function getNameForLogs()
